Move home page navbar to its own file
